Spring boot added Some refactoring
