+++ build.gradle
@@ -1,16 +1,31 @@
-plugins {
-    id 'java'
-    id 'war'
+buildscript {
+    repositories {
+        mavenCentral()
+    }
+    dependencies {
+        classpath("org.springframework.boot:spring-boot-gradle-plugin:2.0.5.RELEASE")
+    }
 }
+
+apply plugin: 'java'
+apply plugin: 'idea'
+apply plugin: 'org.springframework.boot'
+apply plugin: 'io.spring.dependency-management'
 
 group 'ru.innopolis.stc'
 version '1.0-SNAPSHOT'
 
-sourceCompatibility = 1.8
+bootJar {
+    baseName = 'gs-spring-boot'
+    version =  '0.1.0'
+}
 
 repositories {
     mavenCentral()
 }
+
+sourceCompatibility = 1.8
+targetCompatibility = 1.8
 
 test {
     useJUnitPlatform()
@@ -30,6 +45,11 @@ dependencies {
     compile group: 'log4j', name: 'log4j', version: '1.2.16'
     compile group: 'org.postgresql', name: 'postgresql', version: '42.2.5'
     compile group: 'postgresql', name: 'postgresql', version: '9.1-901-1.jdbc4'
-    compile 'javax.servlet:javax.servlet-api:3.1.0'
+//    compile 'javax.servlet:javax.servlet-api:3.1.0'
     compile group: 'jstl', name: 'jstl', version: '1.2'
+    compile("org.springframework.boot:spring-boot-starter-web")
+    compile("org.springframework.boot:spring-boot-starter-thymeleaf")
+    compile("org.springframework.boot:spring-boot-starter-security")
+    compile("org.springframework.boot:spring-boot-devtools")
+    testCompile("junit:junit")
 }
