+++ src/main/java/ru/innopolis/stc/controller/CourseController.java
@@ -1,0 +1,8 @@
+package ru.innopolis.stc.controller;
+
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class CourseController {
+    // Все контроллеры для отображения вьюх связанных с курсами
+}
