+++ src/main/java/ru/innopolis/stc/controller/LessonController.java
@@ -1,0 +1,8 @@
+package ru.innopolis.stc.controller;
+
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class LessonController {
+    // Все контроллеры для отображения вьюх связанных с уроками
+}
