--- src/main/java/ru/innopolis/stc/controller/StartPage.java
@@ -1,18 +1,0 @@
-package ru.innopolis.stc.controller;
-
-import ru.innopolis.stc.pojo.User;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-public class StartPage extends HttpServlet {
-    @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        User user = (User) req.getSession().getAttribute("user");
-        req.setAttribute("user", user);
-        req.getRequestDispatcher("WEB-INF/pages/start_page.jsp").forward(req, resp);
-    }
-}
