+++ src/main/java/ru/innopolis/stc/controller/UserController.java
@@ -1,0 +1,8 @@
+package ru.innopolis.stc.controller;
+
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class UserController {
+    // Все контроллеры для отображения вьюх связанных с управлением пользователями
+}
