+++ src/main/java/ru/innopolis/stc/dao/CourseDao.java
@@ -6,12 +6,8 @@ import java.util.List;
 
 public interface CourseDao {
     Course add(Course course);
-
     Course update(Course course);
-
     Course getById(int id);
-
     Course delete(Course course);
-
     List<Course> getAll();
 }
