+++ src/main/java/ru/innopolis/stc/dao/UserDao.java
@@ -4,12 +4,8 @@ import ru.innopolis.stc.pojo.User;
 
 public interface UserDao {
     User add(User user);
-
     User getById(int id);
-
     User getByEmail(String email);
-
     void update(User user);
-
     void delete(User user);
 }
