+++ src/main/java/ru/innopolis/stc/daosql/CourseDaoImpl.java
@@ -5,7 +5,6 @@ import ru.innopolis.stc.dao.CourseDao;
 import ru.innopolis.stc.db.connectionPool.DatabaseConnectionPool;
 import ru.innopolis.stc.pojo.Course;
 import ru.innopolis.stc.pojo.Teacher;
-
 import java.sql.*;
 import java.util.ArrayList;
 import java.util.List;
