+++ src/main/java/ru/innopolis/stc/utils/ReadPropertiesFile.java
@@ -1,12 +1,11 @@
 package ru.innopolis.stc.utils;
 
 import org.apache.log4j.Logger;
-import ru.innopolis.stc.controller.StartPage;
 
 import java.util.ResourceBundle;
 
 public class ReadPropertiesFile {
-    private final static Logger LOGGER = Logger.getLogger(StartPage.class);
+    private final static Logger LOGGER = Logger.getLogger(ReadPropertiesFile.class);
 
     public String getProperty(String fileName, String propertyName) {
         String property = null;
