+++ src/main/resources/sql.backup
@@ -60,7 +60,7 @@ ALTER TABLE public.course_user OWNER TO postgres;
 CREATE TABLE lesson (
     id integer NOT NULL,
     course_id integer NOT NULL,
-    name character NOT NULL,
+    name integer NOT NULL,
     content character varying,
     home_work character varying
 );
@@ -133,7 +133,7 @@ CREATE TABLE teacher (
     id integer NOT NULL,
     user_id integer,
     description character varying,
-    photo bytea
+    photo character varying
 );
