+++ src/main/resources/templates/pages/all_courses.jsp
