+++ src/main/resources/templates/pages/create-course.jsp
