+++ src/main/resources/templates/pages/registration.jsp
