+++ src/main/resources/templates/pages/start_page.jsp
