--- src/main/webapp/WEB-INF/web.xml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<web-app xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xmlns="http://xmlns.jcp.org/xml/ns/javaee"
-         xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/javaee http://xmlns.jcp.org/xml/ns/javaee/web-app_4_0.xsd"
-         version="4.0">
-
-    <display-name>Servlet</display-name>
-    <servlet>
-        <servlet-name>startPage</servlet-name>
-        <servlet-class>ru.innopolis.stc.controller.StartPage</servlet-class>
-    </servlet>
-
-    <servlet>
-        <servlet-name>allCourses</servlet-name>
-        <servlet-class>ru.innopolis.stc.controller.CourseServlet</servlet-class>
-    </servlet>
-
-    <servlet>
-        <servlet-name>createCourse</servlet-name>
-        <servlet-class>ru.innopolis.stc.controller.CreateCourseServlet</servlet-class>
-    </servlet>
-
-    <!--Маппинги-->
-    <servlet-mapping>
-        <servlet-name>startPage</servlet-name>
-        <url-pattern>/start_page</url-pattern>
-    </servlet-mapping>
-
-    <servlet-mapping>
-        <servlet-name>allCourses</servlet-name>
-        <url-pattern>/courses</url-pattern>
-    </servlet-mapping>
-
-    <servlet-mapping>
-        <servlet-name>createCourse</servlet-name>
-        <url-pattern>/course/create</url-pattern>
-    </servlet-mapping>
-
-    <servlet>
-        <servlet-name>registration</servlet-name>
-        <servlet-class>ru.innopolis.stc.controller.Registration</servlet-class>
-    </servlet>
-
-    <servlet-mapping>
-        <servlet-name>registration</servlet-name>
-        <url-pattern>/registration</url-pattern>
-    </servlet-mapping>
-
-
-</web-app>
